requirements.txt
# Core dependencies for Artifice vs Nature
# PyTorch 1.7.0 and torchvision 0.8.1 installed from .whl in lib/

# Note: numpy 1.19.x compatible with PyTorch 1.7.0 but pandas needs >=1.20.3
# If you have pandas installed, use: numpy>=1.20.3,<1.25.0
# For PyTorch 1.7.0 only: numpy>=1.19.0,<1.20.0
numpy>=1.19.0
scipy>=1.5.0
Pillow>=8.0.0
matplotlib>=3.3.0
pyyaml>=5.3.0
jupyter>=1.0.0
notebook>=6.1.0
